Finished validation, started testing and logging (not working)

+++ backend/RoomPlanner/room-planner/src/main/java/edu/roomplanner/service/UserService.java
@@ -8,6 +8,6 @@ public interface UserService {
 
     List<RoomDto> getAllRooms();
 
-    RoomDto getRoomById(Integer id);
+    RoomDto getRoomById(Long id);
 
 }
